buffet: support using default flag values
A previous change added support for overridding the configuration and
state paths, for testing, but resulted in the defaults not being used
since the empty string was passed for a flag value.  This changes the
init script so that empty paths from the upstart environment are not
used as flag values.

TEST=built, tested on device, tested with override
BUG=none

Change-Id: Id680ec7434b4c523cd0a2c8419dcf342a0a34199
Reviewed-on: https://chromium-review.googlesource.com/248580
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Aaron Kemp <kemp@google.com>
Tested-by: Aaron Kemp <kemp@google.com>