Add weaved process to 'inet' group
In order to be able to open PF_INET/PF_INET6 sockets, the process
must belong to 'inet' group because Brillo's kernel is compiled with
CONFIG_ANDROID_PARANOID_NETWORK=y and this essentially kills requests
to socket(2) when creating outbound sockets. This prevents HTTP(S)
requests to go out to the cloud server and stops device registration
process.

BUG:23416684
Change-Id: I98767b97b44b194ef22a0036e93d9ef90320c484