Don't fall-back into AP mode after connecting to WiFi network
Probably the code was written with assumption that
On*PropertyChangeRegistration will be called asynchronously.
These function access data which will be initialized immediately after
RegisterPropertyChangedSignalHandler return.
This is not true with the current state of the code. Functions are being
called from RegisterPropertyChangedSignalHandler before data is read.
This effectively disabled most of OnPropertyChange logic. Probably the
logic of DBUS proxies was changed.

This fix initializes required data before calling
RegisterPropertyChangedSignalHandler.

Also fixed call to Init(), but seems does nothing important.

BUG:25531670
Change-Id: I7e1580c165d618b3f9bd763d06f909d9c35f87c3