weaved: set system property 'weave.state' with the current GCD state
In order for other Brillo system components to know when the device is
fully provisioned/connected to the cloud, add 'weave.state' system
property which has one of the following states:

- unconfigured
- connecting
- connected
- invalid_credentials

Initially (OOB) the device is "unconfigured". When the bootstrap
process finished, the device will be in "connected" state. If the
device is later deleted from the cloud by the user, it will go into
"invalid_credentials" state.

BUG: 24142361
Change-Id: If83e151940a81d5e379e1b0199a9813733c12fdf