Fix avahi host name resolution scenarios
When host name collision is detected, avahi picks a new host name,
however it fails to update any published services with the new name
which leads to wrong IP address resolution for the service.

As a work-around, when we detect an avahi-client restart, we just
re-publish our service, so it picks up the new host name.

BUG: 26237307
Change-Id: I980735890c62957e1bc433a07c319eff646c358c