weaved: Fix binder connection logic from client side
When weaved client tries to connect to weaved process over binder,
it first obtains the binder service exposed by weaved, then invokes
connect() method and provides a callback binder object which is
invoked when weaved is fully initialized and connected.

Only then the client monitors the provided IWeaveService for death
notifications. However if weaved crashes during this process, the
client will never notice the death of weaved.

This change makes the client to subscribe for death notifications
as soon as IWeaveServiceManager binder is obtained instead of using
IWeaveService provided in IWeaveClient::onServiceConnected()

BUG: 26981732
Change-Id: I9b0c3b020558c336a933bdfe106f73673294c062